projects/beneficios-amba/backend_api/requirements.txt
django>=5.0.0
djangorestframework
pillow
django-cors-headers
psycopg2-binary
dj-database-url
